Changing edit page to an edit modal
